+++ forui/lib/src/widgets/text_field/text_field.dart
@@ -5,13 +5,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:meta/meta.dart';
 
 import 'package:forui/forui.dart';
-
-part 'text_field_state_style.dart';
-part 'text_field_style.dart';
-part 'text_form_field.dart';
+import 'package:forui/src/widgets/text_field/field.dart';
 
 /// A text field.
 ///
@@ -738,7 +734,7 @@ final class FTextField extends StatelessWidget {
               primaryColor: style.cursorColor,
             ),
           ),
-          child: _Field(
+          child: Field(
             parent: this,
             style: style,
             key: key,